combine_types: Use BTreeSet for path list
Doing this means we take O(1) insertion operations followed by a sort
operation that is O(n log n) / O(n) and turn it into a series of O(log n)
insertion operations.

The benefit here is that BTreeSet is always sorted as long as the Ord
implementation is stable (which it is in this case).  So we can just
insert all the items and be done.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>