handlers: Add bonapp module with url encoding example
To get started with, url encoding is required to pass the ?cafe= part
of the request URI to the BonApp API.  Having this done via serde seems
to work well enough. We can always hard code later or see if there is an
upstream way to do this that doesn't involve invoking serde every time.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>