bonapp: Return single Cafe response rather than CafesResponse
In the reference implementation, this is something we do, where we
transform from:

    {
      "cafes": {
        "123": {
          /* cafe response */
        }
      }
    }

to just

    {
      /* cafe response */
    }

by taking the .cafes.<id> part and discarding the rest.  Note that we
explicitly _delete_ the remaining contents of the response structure
in JavaScript-land (probably to avoid lots of garbage to collect or
something of that sort) but this is not necessary here.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>