Include lint configuration in all crates
Rather than relying on a big long special command getting run, we can
actually inline this configuration by using an _inner_ attribute on each
of the targets.  Note that `src/main.rs` is the target for the server,
whereas `handlers`, `routes`, and `types` each have a `src/lib.rs` where
this is applied instead.

This should produce identical results to what we got with the special CI
configuration.  Now it's tracked in the code, though.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>