bonapp/menu: Handle transforming multi-cafe response to single-cafe
In the reference implementation, we have a block

    let days = resp.body.days.map((dayInfo) => {
        let info = Object.assign({}, dayInfo)
        info.cafe = info.cafes[cafeId]
        delete info.cafes
        return info
    })

which transforms the days key to not have multiple entries per cafe.
Since our API currently only queries for a single cafe's information, it
makes sense to apply this transformation.

I think I ultimately want to turn the BonAppMenu{Single,Multiple}...
type into an enum and use #[serde(untagged)] to treat them as a single
type.  Then, the conversion _could_ be smarter (e.g., if there are
multiple cafes, leave the response as-is).

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>