e2e: Try to parse body to serde_json::Value
Parsing and then reserializing is an important step because
serde_json::Value (specifically, the Object variant) is sort-stable. By
default, the backing storage is a BTreeMap. It's not as much of a
concern with our server since it uses serde_json already, but since the
reference servers use JavaScript, we might get equivalent JSON that just
lists things in a different field-order. First parsing, then reprinting
(using pretty-print) ensures that the differences we should get are
actual "real" differences in the object structure or values.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>