switch ts-rs dependency to frogpond fork
Beta clippy recently started failing on a downstream consumer (issue) of this
crate. This was because of a number of calls to vec! that were picked up by
clippy. This lint only fires, at time of writing, in beta (stable does not have
the change that introduced it, and nightly has since changed to ignore this
lint in macros...). This will, assuming this change gets released quickly, help
fellow #![deny(clippy::all)] folks who use this crate.

Also removes some trailing whitespace in ts-rs/src/lib.rs in d4dfefe

Co-authored-by: Kristofer Rye <kristofer.rye@gmail.com>