keep default case but opt to use custom formatter for default structure
We prefer fmt+init to retain a compact ouput whereas layer+boxed is overly
verbose. Since the registry is still appropriate in all cases this helps us
achieve the same concise output without the registry call (and keeps the Default
case). The registry was driving all of the additional context and verbosity
behind the span traces.