workspace: Configure deny-default lint precedence
One key difference when housing lint configuration in Cargo.toml is that
the order is ignored. For a lint group such as clippy::all, this has the
same priority as a lint.  If the lint is included in this group and then
fires, this creates a problem when determining the lint level: if the
lint says its default is "warn" but the group says "deny", the level is
ambiguous.

Instead, mark the `level = deny` lint as having a lower priority, such
that if a more specific rule is defined it takes precedence.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>