Resolve a couple of obfuscation clippy warnings
With Rust 1.87 beta, clippy started warning about some new instances of
the obfuscated-if-else lint. This appears to have been introduced in
rust-lang/rust-clippy#14165, which was a change that made sense to me.

But because we deny the warn-by-default group, clippy::all, CI began
failing, which is fine.

This commit resolves the two instances of this lint in the codebase by
using inline if-else's.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>